requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
openai==1.51.0
python-multipart==0.0.9
httpx==0.27.2
cloudinary==1.40.0
pandas==2.1.4
requests==2.32.3
openpyxl>=3.1.5
pymupdf==1.24.10
pdfplumber==0.11.4
playwright==1.47.0
pytesseract==0.3.10
opencv-python-headless==4.10.0.84
google-generativeai==0.8.6
beautifulsoup4==4.14.3
sqlmodel
asyncpg
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
